Change command calling convention
In order to have more flexible interface a git-cli-like behaviour is
employed: when a user passes a command to `ds` the utility looks for an
executable on PATH with the "ds-<command>" name. If the executable is
found it gets to be started with passed parameters otherwise ds reports
that it cannot execute the command. All main scripts have been renamed
to follow the calling convention.

Signed-off-by: Nikita Eshkeev <neshkeev@yandex.ru>