[ds-pull] Change processing of ds-pull options
This patch changes the processing of ds-pull options. Now a user can set
either "--url" or "container_name" and ds-pull will reject if two are
used simultaneously. If only "--url" is specified, then "container_name"
is deduced from its value.

If the user set options that are not part of the ds-pull's interface
then they are passed to curl, hence the user can manupulate headers,
authentications and other stuff properly.

Signed-off-by: Nikita Eshkeev <neshkeev@yandex.ru>