[ds-create-ubuntu] Add script to create ubuntu containers
This patch adds a new script that creates ubuntu containers with
debootstrap.

The possibility to adapt the script to create Debian containers needs to
be explored.

Also there are duplicating functions in ds-create-ubuntu and
ds-create-arch. They need to be extracted into either "util.sh" or
"create-util.sh" or something similar.

Signed-off-by: Nikita Eshkeev <neshkeev@yandex.ru>